backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.30.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
pydantic==2.9.0
pydantic-settings==2.5.0
anthropic==0.39.0
httpx==0.27.0
python-dotenv==1.0.1
